extract test for shape collection
